shared show view kinda working!! nice!!! need to think about how to trigger refresh on the page when executing commands from shared view, but this is not crucial -- added swipe down to refresh on today's view for now
