+++ ListeningCompanion/SearchView.xaml.cs
@@ -17,9 +17,6 @@ public partial class SearchView : ContentPage
 {
     #region Fields
     private const string connectionString = @"Server=tcp:listeningcompanion.database.windows.net,1433;Initial Catalog=ListeningCompanion;Persist Security Info=False;User ID=captaintrips;Password=TerrapinStation77!;MultipleActiveResultSets=False;Encrypt=True;TrustServerCertificate=False;Connection Timeout=30;";
-
-    public ICommand BookmarkCommand { get; private set; }
-    public ICommand LikeCommand { get; private set; }
     public ICommand ShowSelectedCommand { get; private set; }
     public List<UserShowDetails> showResults { get; private set; } = new List<UserShowDetails>();
     CollectionView showResultsView { get; set; }
@@ -165,63 +162,8 @@ public partial class SearchView : ContentPage
 
     public async Task<CollectionView> LoadShowCollectionView()
     {
-        CollectionView showsCollectionView = new CollectionView();
-        showsCollectionView.ItemsSource = showResults;
-        //showsCollectionView.SetBinding(ItemsView.ItemsSourceProperty, "UserShowDetails");
-        showsCollectionView.ItemTemplate = new DataTemplate(() =>
-        {
-            Grid grid = new Grid { Padding = 10 };
-            grid.RowDefinitions.Add(new RowDefinition { Height = GridLength.Auto });
-            grid.RowDefinitions.Add(new RowDefinition { Height = GridLength.Auto });
-            grid.ColumnDefinitions.Add(new ColumnDefinition { Width = GridLength.Auto });
-            grid.ColumnDefinitions.Add(new ColumnDefinition { Width = GridLength.Auto });
-
-
-            Label showDateLabel = new Label { FontAttributes = FontAttributes.Bold };
-            showDateLabel.SetBinding(Label.TextProperty, "Date");
-
-            Label venueLabel = new Label { FontAttributes = FontAttributes.Italic, VerticalOptions = LayoutOptions.End };
-            venueLabel.SetBinding(Label.TextProperty, "VenueName");
-
-            Label statusLabel = new Label { FontAttributes = FontAttributes.Italic, VerticalOptions = LayoutOptions.End };
-            statusLabel.SetBinding(Label.TextProperty, "InteractionStatus");
-
-            grid.Add(showDateLabel, 0, 0);
-            grid.Add(venueLabel, 0, 1);
-            grid.Add(statusLabel, 1, 0);
-
-
-            SwipeView swipeView = new SwipeView();
-            SwipeItem bookmarkSwipeItem = new SwipeItem
-            {
-                Text = "Bookmark",
-                IconImageSource = ImageSource.FromFile("bookmark_big.png"),
-                BackgroundColor = Colors.Green,
-                Command = BookmarkCommand
-            };
-            //bookmarkSwipeItem.SetBinding(MenuItem.CommandProperty, new Binding("BindingContext.BookmarkCommand", source: showsCollectionView));
-            bookmarkSwipeItem.SetBinding(MenuItem.CommandParameterProperty, ".");
-
-            SwipeItem likeSwipeItem = new SwipeItem
-            {
-                Text = "Like",
-                IconImageSource = ImageSource.FromFile("star_big.png"),
-                BackgroundColor = Colors.GreenYellow,
-                Command = LikeCommand
-            };
-            likeSwipeItem.SetBinding(MenuItem.CommandParameterProperty, ".");
-
-
-            swipeView.RightItems = new SwipeItems { bookmarkSwipeItem, likeSwipeItem };
-
-
-
-            swipeView.Content = grid;
-
-            return swipeView;
-        });
-        showsCollectionView.SelectionMode = SelectionMode.Single;
         // Define SelectionChanged event handler
+        CollectionView showsCollectionView = new ShowCollectionView(connectionString).GetCollectionViewFromUserShowDetailsList(showResults);
         showsCollectionView.SelectionChanged += async (sender, e) =>
         {
             if (e.CurrentSelection.FirstOrDefault() is UserShowDetails selectedItem)
@@ -234,47 +176,6 @@ public partial class SearchView : ContentPage
     #endregion
 
     #region Commands
-
-    private async void ExecuteBookmarkCommand(object song)
-    {
-        var selectedSong = (UserSongDetails)song;
-        UserPerformedSong userPerformedSong = new UserPerformedSong();
-        if (selectedSong.UserPerformedSongId > 0)
-        {
-            userPerformedSong.ID = selectedSong.UserPerformedSongId;
-        }
-        userPerformedSong.Bookmarked = !selectedSong.SongBookmarked;
-        userPerformedSong.PerformedSongID = selectedSong.PerformedSongId;
-        userPerformedSong.Liked = selectedSong.SongLiked;
-        userPerformedSong.Rating = selectedSong.SongRating;
-        userPerformedSong.Notes = selectedSong.SongNotes;
-        userPerformedSong.Rating = selectedSong.SongRating;
-        userPerformedSong.UserID = 1;
-        UserPerformedSongService userPerformedSongService = new UserPerformedSongService(connectionString);
-        userPerformedSongService.SaveUserPerformedSong(userPerformedSong);
-        LoadSearchPage(true);
-
-    }
-    private async void ExecuteLikeCommand(object song)
-    {
-        var selectedSong = (UserSongDetails)song;
-        UserPerformedSong userPerformedSong = new UserPerformedSong();
-        if (selectedSong.UserPerformedSongId > 0)
-        {
-            userPerformedSong.ID = selectedSong.UserPerformedSongId;
-        }
-        userPerformedSong.Bookmarked = selectedSong.SongBookmarked;
-        userPerformedSong.PerformedSongID = selectedSong.PerformedSongId;
-        userPerformedSong.Liked = !selectedSong.SongLiked;
-        userPerformedSong.Rating = selectedSong.SongRating;
-        userPerformedSong.Notes = selectedSong.SongNotes;
-        userPerformedSong.Rating = selectedSong.SongRating;
-        userPerformedSong.UserID = 1;
-        UserPerformedSongService userPerformedSongService = new UserPerformedSongService(connectionString);
-        userPerformedSongService.SaveUserPerformedSong(userPerformedSong);
-        LoadSearchPage(true);
-
-    }
     public void OnFilterRangeChanged(object sender, CheckedChangedEventArgs e)
     {
         if (filterRange.IsChecked)
